requirements.txt
# MRAS v3.0 — Pinned Python dependencies

# ---- Web framework ----
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ---- ORM / Database ----
sqlmodel==0.0.21
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# ---- Settings ----
pydantic-settings==2.5.2

# ---- Auth ----
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ---- ML / Forecasting ----
scikit-learn==1.5.2
prophet==1.1.5
pandas==2.2.3
numpy==1.26.4

# ---- LLM ----
anthropic==0.34.2

# ---- Notifications ----
twilio==9.2.4
sendgrid==6.11.0

# ---- HTTP client ----
httpx==0.27.2

# ---- Scheduler ----
apscheduler==3.10.4

# ---- Testing ----
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# ---- Code quality ----
ruff==0.6.9
mypy==1.11.2
